requirements.txt
Django>=3.0,<3.2
psycopg2-binary>=2.8
requests>=2.25
djangorestframework==3.12.4
markdown>=3.3
django-filter==2.4